read notifications and post Toast with title
This base solution for getting notification is worked, when app is not in recent and in recent too.
In future, we can filter by package id and parse notification of banks locally, show it in time line. For Tinkoff we can get product name (mTele2, for example)

On binary changes in app we should request and re-enable permission